Create new branch and reorganize repository organization
