+++ src/Helpers/fileFormatCheck.py
@@ -1,5 +1,5 @@
 import os
-from config import ALLOWED_FORMATS
+from src.Settings.config import ALLOWED_FORMATS
 
 class InvalidFileFormatError(Exception):
     """Custom exception for invalid file formats."""
